+++ analysis/import/GitLogParser/src/test/kotlin/de/maibornwolff/codecharta/importer/gitlogparser/GitLogParserTest.kt
@@ -1,6 +1,8 @@
 package de.maibornwolff.codecharta.importer.gitlogparser
 
 import de.maibornwolff.codecharta.importer.gitlogparser.GitLogParser.Companion.main
+import de.maibornwolff.codecharta.serialization.ProjectDeserializer
+import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import java.io.File
@@ -13,13 +15,13 @@ class GitLogParserTest {
             arrayOf(
                 "repo-scan",
                 "--repo-path=../../..",
-                "--output-file=src/test/resources/gitlog-analysis.cc.json",
+                "--output-file=src/test/resources/gitlog-analysis-repo.cc.json",
                 "--not-compressed",
                 "--silent=false",
                 "--add-author=false"
             )
         )
-        val file = File("src/test/resources/gitlog-analysis.cc.json")
+        val file = File("src/test/resources/gitlog-analysis-repo.cc.json")
         file.deleteOnExit()
 
         assertTrue(file.exists())
@@ -32,15 +34,16 @@ class GitLogParserTest {
                 "log-scan",
                 "--git-log=src/test/resources/codeCharta.log",
                 "--repo-files=src/test/resources/names-in-git-repo.txt",
-                "--output-file=src/test/resources/gitlog-analysis.cc.json",
+                "--output-file=src/test/resources/gitlog-analysis-log.cc.json",
                 "--not-compressed",
                 "--silent=true"
             )
         )
-        val file = File("src/test/resources/gitlog-analysis.cc.json")
+        val file = File("src/test/resources/gitlog-analysis-log.cc.json")
         file.deleteOnExit()
-
         assertTrue(file.exists())
+        val project = ProjectDeserializer.deserializeProject(file.inputStream())
+        assertEquals(project.attributeDescriptors, getAttributeDescriptors())
     }
 
     @Test